Adding dummy data that is automaticly added when running the db setup

+++ Domain/Models/Sensor.cs
@@ -5,8 +5,8 @@ public class Sensor
     public int Id { get; set; }
     public int FieldId { get; set; }
     public Field Field { get; private set; }
-    public int Longitude { get; set; }
-    public int Latitude { get; set; }
+    public double Longitude { get; set; }
+    public double Latitude { get; set; }
     public int MoistureLevel { get; set; }
     public int soiltype { get; set; }
 }
